記事取得を共通化
