+++ src/app/posts/page/[pageNum]/page.tsx
@@ -1,35 +1,14 @@
-import matter from "gray-matter";
 import { Posts } from "../../../../components/Posts";
 import { PostsPaginate } from "../../../../components/PostsPaginate";
 import { postsPerPage } from "../../../../utils/constants";
-import { PostData } from "../../../../utils/posts";
-type Value = {
-  default: string;
-};
+import { getAllPosts } from "../../getAllPosts";
 
 type Params = {
   pageNum?: string;
 };
 
 const PostsPage = async ({ params }: { params: Params }) => {
-  const allPosts = ((context) => {
-    const keys = context.keys().filter((fileName) => /^\.\//.test(fileName));
-    const values = keys.map(context);
-    const data = keys.map((key, index) => {
-      const slug = key.replace(/^.*[\\/]/, "").slice(0, -3);
-      const value = values[index] as Value;
-      const meta = matter(value.default);
-      return {
-        type: "markdown",
-        metaData: meta.data,
-        slug,
-        href: `/posts/${slug}`,
-      };
-    });
-    return data;
-  })(
-    require.context("../../../../../data/posts", true, /\.*\.md$/)
-  ) as PostData[];
+  const allPosts = getAllPosts();
   const totalPage = Math.ceil(allPosts.length / postsPerPage);
   const pageNum = parseInt(params?.pageNum || "1", 10);
   const start = (pageNum - 1) * postsPerPage;
